Move menu/pickers to onChange event and send onSelect from MenuItem
Added a second event for item selection since the mix of inheritance
and the decorator pattern made using onSelect for everything too
complicated. Now onChange is used rather than onSelect by the picker
kinds to communicate changes between the activator (PickerButton),
PickerDecorator and Picker. onSelect is used by MenuItem to communicate
to the client when a menu choice is made.

Note that FlyweightPicker stops the MenuItem onSelect event from
reaching the client since it will not have the correct value in a
flyweight context. FlyweightPicker sends its own onSelect event once it
has the correct selected item value to send.