Update for ENYO-518
Reposition tooltips above their decorator if their bottom is below the
window height & to the left of the decorator if their right side goes
off the right window side. Uses a css strategy for placing the tooltip
arrows on the left, right, top or bottom.

Not 100% perfect in that a window resized just right can cut off a
small piece of the tooltip on any side, however this should rarely be
the case and mostly for decorators that are either really small (past
recommended touch sizes) or off the window, both of which would seem to
be a bigger UI concern. The benefit of the current strategy is that it
allows for a clean and 'simple' implementation without doing a large
amount of position processing in JS or using additional elements for
the tooltip arrows.

Also added documentation.