Improved menu positioning
Added re-positioning calculation for when a menu gets cut off by the
right side of the screen to move it over to the left by the proper
amount to put it fully back in view.

Also added a check for vertical positioning to make sure that before a
menu is moved above an activator (when it is being cut off at the
bottom of the view) that there is actually more room than below.