Temporary fix for ENYO-674
Forcing toolbars to re-render when any contained menus are hidden. See
bug notes for more info but basically there are certain conditions that
cause a component to be hidden & toolbars with menus are a common
situation this arises so this fix addresses them specifically. Future
fix will address root cause.