Added onChange event to onyx.MenuItem.
When the content for a MenuItem changed it was not letting any of it's parents know about the change.  Consequently when MenuItem was used inside of a PickerDecorator then the PickerDecorator (and more importantly the PickerButton child component) was never notified when a MenuItems content changed.

That creates the case where you have a PickerDecorator on screen and the selected MenuItems content changes but the PickerButtons content is not updated with the change.