Fixed bug that caused MenuItem content changes to trigger PickerButton content changes even when the MenuItem was not the currently selected MenuItem.
The problem with the previous commit was that the change event was being used to notify the PickerButton that the selected item had changed.  Before that was fine since MenuItem didn't fire a change event even when it's content changed.

Since PickerDecorator was relying on change events to only occur when the selected item changed it caused the bug when I added the change event to MenuItem.

**In this commit I've changed the event handling to be:**

MenuItem - Fires change event when content changes.

Picker - Handles change events.  If the change event originates from the currently selected item the Picker will fire a selectedContentChange event.

PickerDecorator - Handles selectedContentChange events by waterfallDown.

PickerButton - Handles selectedContentChange events by updating the buttons content.